Webpack included to add import options for js files and make all libraries work together easier.
