+++ contactme.html
@@ -6,25 +6,21 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="icon" type="image/svg+xml" href="./public/img/favico.jpg" />
 
-    <link rel="stylesheet" href="./css/sideBar.css" />
-    <link rel="stylesheet" href="./css/style.css" />
-    <link rel="stylesheet" href="./css/header.css" />
-    <link rel="stylesheet" href="./css/main.css" />
-    <link rel="stylesheet" href="./css/footer.css" />
-    <link rel="stylesheet" href="./css/menu.css">
+    <link rel="stylesheet" href="./src/css/sideBar.css" />
+    <link rel="stylesheet" href="./src/css/style.css" />
+    <link rel="stylesheet" href="./src/css/header.css" />
+    <link rel="stylesheet" href="./src/css/main.css" />
+    <link rel="stylesheet" href="./src/css/footer.css" />
+    <link rel="stylesheet" href="./src/css/menu.css">
         
     <title>Skrollu | Contact me</title>
 </head>
 <body>
     <main id="swup">
         <h1>Contact me</h1>
-        <a href="/portfolio/menu.html">Menu</a>
+        <a href="/menu.html">Menu</a>
     </main>
     
-    <script src="./js/libraries/swup.min.js"></script>
-    <script src="./js/libraries/SwupFadeTheme.min.js"></script>
-    <script src="./js/libraries/SwupHeadPlugin.min.js"></script>
-    
-    <script src="./js/transition.js"></script>
+    <script  src="./dist/js/bundle.js"></script>
 </body>
 </html>
