+++ menu.html
@@ -10,12 +10,12 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@900&display=swap" rel="stylesheet"> 
     
-    <link rel="stylesheet" href="./css/sideBar.css" />
-    <link rel="stylesheet" href="./css/style.css" />
-    <link rel="stylesheet" href="./css/header.css" />
-    <link rel="stylesheet" href="./css/main.css" />
-    <link rel="stylesheet" href="./css/footer.css" />
-    <link rel="stylesheet" href="./css/menu.css">
+    <link rel="stylesheet" href="./src/css/sideBar.css" />
+    <link rel="stylesheet" href="./src/css/style.css" />
+    <link rel="stylesheet" href="./src/css/header.css" />
+    <link rel="stylesheet" href="./src/css/main.css" />
+    <link rel="stylesheet" href="./src/css/footer.css" />
+    <link rel="stylesheet" href="./src/css/menu.css">
         
     <title>Skrollu | Menu</title>
 </head>
@@ -23,16 +23,12 @@
     
     <nav id="swup">
         <ul>
-            <li><a class="link" href="/portfolio/projects.html">Portfolio</a></li>
-            <li><a class="link" href="/portfolio/aboutme.html">About me</a></li>
-            <li><a class="link" href="/portfolio/contactme.html">Contacts</a></li>
+            <li><a class="link" href="/projects.html">Portfolio</a></li>
+            <li><a class="link" href="/aboutme.html">About me</a></li>
+            <li><a class="link" href="/contactme.html">Contacts</a></li>
         </ul>
     </nav>
     
-    <script src="./js/libraries/swup.min.js"></script>
-    <script src="./js/libraries/SwupFadeTheme.min.js"></script>
-    <script src="./js/libraries/SwupHeadPlugin.min.js"></script>
-
-    <script src="./js/transition.js"></script>
+    <script  src="./dist/js/bundle.js"></script>
 </body>
 </html>
