+++ projects.html
@@ -5,12 +5,12 @@
   <link rel="icon" type="image/svg+xml" href="./public/img/favico.jpg" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   
-    <link rel="stylesheet" href="./css/sideBar.css" />
-    <link rel="stylesheet" href="./css/style.css" />
-    <link rel="stylesheet" href="./css/header.css" />
-    <link rel="stylesheet" href="./css/main.css" />
-    <link rel="stylesheet" href="./css/footer.css" />
-    <link rel="stylesheet" href="./css/menu.css">
+    <link rel="stylesheet" href="./src/css/sideBar.css" />
+    <link rel="stylesheet" href="./src/css/style.css" />
+    <link rel="stylesheet" href="./src/css/header.css" />
+    <link rel="stylesheet" href="./src/css/main.css" />
+    <link rel="stylesheet" href="./src/css/footer.css" />
+    <link rel="stylesheet" href="./src/css/menu.css">
   
   <link rel="preconnect" href="https://fonts.gstatic.com">
   <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600&display=swap" rel="stylesheet">
@@ -37,7 +37,7 @@
   
 <main id="swup">
   
-  <a id="menu-btn" href="/portfolio/menu.html">
+  <a id="menu-btn" href="/menu.html">
     <img class="hamburger-img" src="./public/img/hamburger.png" alt="Go to menu">
   </a>
   
@@ -197,22 +197,7 @@
     </footer>
   </main>
     
-  <script type="module" src="/js/header.js"></script>
-  <!--
-    <script src="js/sideBar.js"></script>
-    <script type="module" src="js/main.js"></script>
-  -->
-
-<!-- SWUP -->
-  <script src="./js/libraries/swup.min.js"></script>
-  <script src="./js/libraries/SwupFadeTheme.min.js"></script>
-  <script src="./js/libraries/SwupHeadPlugin.min.js"></script>
-<!-- GSAP -->
-  <script src="./js/libraries/gsap.min.js"></script>
-  <script src="./js/libraries/TextPlugin.min.js"></script>
-  <script src="./js/libraries/ScrollTrigger.min.js"></script>
-
-  <script  src="./js/transition.js"></script>
+  <script  src="./dist/js/bundle.js"></script>
   
 </body>
 </html>
